Move macros for range computations into a separate file
It's tidier.
