+++ src/types.rs
@@ -38,125 +38,6 @@ const META_BLOCK_SIZE: Sectors = Sectors(8);
 #[allow(dead_code)]
 const MAX_META_DEV_SIZE: MetaBlocks = MetaBlocks(255 * ((1 << 14) - 64));
 
-// division by self
-macro_rules! self_div {
-    ($T:ident) => {
-        impl Div<$T> for $T {
-            type Output = u64;
-            fn div(self, rhs: $T) -> u64 {
-                self.0 / *rhs
-            }
-        }
-    };
-}
-
-// implement Debug in a more compact style, e.g. "Sectors(0)"
-macro_rules! debug {
-    ($T:ident) => {
-        impl fmt::Debug for $T {
-            fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-                write!(f, stringify!($T))?;
-                write!(f, "({})", **self)
-            }
-        }
-    };
-}
-
-// macros for implementing serialize and deserialize on all types
-macro_rules! serde {
-    ($T:ident) => {
-        impl serde::Serialize for $T {
-            fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-            where
-                S: serde::Serializer,
-            {
-                serializer.serialize_u64(**self)
-            }
-        }
-
-        impl<'de> serde::Deserialize<'de> for $T {
-            fn deserialize<D>(deserializer: D) -> Result<$T, D::Error>
-            where
-                D: serde::de::Deserializer<'de>,
-            {
-                Ok($T(serde::Deserialize::deserialize(deserializer)?))
-            }
-        }
-    };
-}
-
-// macros for implementing Sum on all types
-macro_rules! sum {
-    ($T:ident) => {
-        impl Sum for $T {
-            fn sum<I: Iterator<Item = $T>>(iter: I) -> $T {
-                iter.fold($T::default(), Add::add)
-            }
-        }
-    };
-}
-
-// macros for unsigned operations on Sectors and Bytes
-macro_rules! unsigned_div {
-    ($t:ty, $T:ident) => {
-        impl Div<$t> for $T {
-            type Output = $T;
-            fn div(self, rhs: $t) -> $T {
-                $T(self.0 / rhs as u64)
-            }
-        }
-    };
-}
-
-macro_rules! unsigned_mul {
-    ($t:ty, $T:ident) => {
-        impl Mul<$t> for $T {
-            type Output = $T;
-            fn mul(self, rhs: $t) -> $T {
-                $T(self.0 * rhs as u64)
-            }
-        }
-
-        impl Mul<$T> for $t {
-            type Output = $T;
-            fn mul(self, rhs: $T) -> $T {
-                $T(self as u64 * rhs.0)
-            }
-        }
-    };
-}
-
-macro_rules! unsigned_rem {
-    ($t:ty, $T:ident) => {
-        impl Rem<$t> for $T {
-            type Output = $T;
-            fn rem(self, rhs: $t) -> $T {
-                $T(self.0 % rhs as u64)
-            }
-        }
-    };
-}
-
-macro_rules! rem {
-    ($T:ident) => {
-        impl Rem<$T> for $T {
-            type Output = $T;
-            fn rem(self, rhs: $T) -> $T {
-                $T(self.0 % rhs.0 as u64)
-            }
-        }
-    };
-}
-
-macro_rules! checked_add {
-    ($T: ident) => {
-        /// Add two items of type $T, return None if overflow.
-        pub fn checked_add(&self, other: $T) -> Option<$T> {
-            self.0.checked_add(other.0).map($T)
-        }
-    }
-}
-
 macro_attr! {
     #[derive(Clone, Copy, Default, Eq, Ord, PartialEq, PartialOrd)]
     /// A type for Data Blocks as used by the thin pool.
